+++ Editor/EditorInternals/InternalsVisibleTo.cs
@@ -2,4 +2,5 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Latios.Core.Editor")]
 [assembly: InternalsVisibleTo("Latios.Kinemation.Editor")]
+[assembly: InternalsVisibleTo("Latios.Transforms.Editor")]