+++ Editor/UIInternals/InternalsVisibleTo.cs
@@ -1,0 +1,6 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Latios.Core.Editor")]
+[assembly: InternalsVisibleTo("Latios.Kinemation.Editor")]
+[assembly: InternalsVisibleTo("Latios.Transforms.Editor")]
+