+++ Kinemation/Systems/SyncPoint/AddMissingMasksSystem.cs
@@ -17,15 +17,15 @@ namespace Latios.Kinemation.Systems
         public void OnCreate(ref SystemState state)
         {
             m_computeQuery = state.Fluent().WithAll<MaterialMeshInfo>(true).WithAll<ComputeDeformShaderIndex>().Without<ShareSkinFromEntity>()
-                             .Without<ChunkComputeDeformMemoryMetadata>(      true).IncludePrefabs().IncludeDisabled().Build();
+                             .Without<ChunkComputeDeformMemoryMetadata>(      true).IncludePrefabs().IncludeDisabledEntities().Build();
 
             m_linearBlendQuery = state.Fluent().WithAll<MaterialMeshInfo>(true).WithAll<LinearBlendSkinningShaderIndex>().Without<ShareSkinFromEntity>()
-                                 .Without<ChunkLinearBlendSkinningMemoryMetadata>(true).IncludePrefabs().IncludeDisabled().Build();
+                                 .Without<ChunkLinearBlendSkinningMemoryMetadata>(true).IncludePrefabs().IncludeDisabledEntities().Build();
 
             m_copyQuery = state.Fluent().WithAll<MaterialMeshInfo>(true).WithAll<ShareSkinFromEntity>(true).Without<ChunkCopySkinShaderData>(true)
-                          .IncludePrefabs().IncludeDisabled().Build();
+                          .IncludePrefabs().IncludeDisabledEntities().Build();
 
-            m_baseQuery = state.Fluent().WithAll<MaterialMeshInfo>(true).Without<ChunkPerFrameCullingMask>(true).IncludePrefabs().IncludeDisabled().Build();
+            m_baseQuery = state.Fluent().WithAll<MaterialMeshInfo>(true).Without<ChunkPerFrameCullingMask>(true).IncludePrefabs().IncludeDisabledEntities().Build();
         }
 
         [BurstCompile]