+++ Documentation~/Kinemation Animation and Rendering/CHANGELOG.md
@@ -6,6 +6,25 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.4] – 2022-12-29
+
+Officially supports Entities [1.0.0 prerelease 15]
+
+### Fixed
+
+-   Fixed exported bones not being assigned correctly during baking when there
+    were more than one in a skeleton
+-   Fixed children of exported bone Game Objects being reparented directly to
+    the skeleton during baking rather than the exported bones
+-   Fixed an issue where Kinemation `SuperSystems` may have been sorted by
+    attribute when such sorting was not desired
+
+### Improved
+
+-   If for some reason an exported bone is not assigned an index or assigned an
+    index to a root, it is now no longer treated as an exported bone and baking
+    will restore default transform components for it
+
 ## [0.6.2] – 2022-12-10
 
 Officially supports Entities [1.0.0 prerelease 15]